Split out builders as modules and implement Will's ES6 changes
